bootstrap: berkeley: add dirs

+++ bootstrap/berkeley.sh
@@ -23,9 +23,11 @@ fi
 set -e
 set -v
 
-mkdir -p ~/.config
-mkdir -p ~/.local/opt
-mkdir -p ~/work
+cd ~
+
+mkdir -p .config
+mkdir -p .local/opt
+mkdir -p work
 
 cd ~/.local/opt
 2> /dev/null git clone https://github.com/szhu/berkeley-cs-build.git
@@ -40,6 +42,7 @@ if [ ! -e ~/.local/bin/fish ]; then
     curses/build.sh
     fish/build.sh
 fi
+rm -rf ~/.local/opt/berkeley-cs-build
 
 cd ~
 fish_bootstrap_cmd='
@@ -50,15 +53,67 @@ if [ -e ~/.local/bin/fish ] && [ "$(arch)" = x86_64 ] && [ -t 1 ] && [ "$SHLVL" 
 fi
 '
 
-for rc_file in ~/.bashrc ~/.bash_profile; do
+fish_bootstrap_cmd='
+if [ -e ~/.local/bin/fish ] && [ "$(arch)" = x86_64 ] && [ -t 1 ] && [ "$SHLVL" -eq 1 ]; then
+    clear
+    ~/.local/bin/fish
+    exit
+fi
+'
+
+for rc_file in .bashrc .bash_profile; do
     if grep -Fq '~/.local/bin/fish' "$rc_file"; then
         echo "$rc_file already invokes fish shell; will not modify"
     else
         cp -f "$rc_file" "$rc_file~"
-        echo "$fish_bootstrap_cmd" >> ~/.bashrc
+        echo "$fish_bootstrap_cmd" >> "$rc_file"
     fi
 done
 
-touch ~/.bootstrapped
+rmdir -f 
 
-~/.local/bin/fish
+rmdir_if_exists() {
+    if [[ -d "$1" ]] ; then
+        rmdir "$1"
+    else
+        echo "$1 gone; will not modify"
+    fi
+}
+mvdir_if_exists() {
+    if [[ -d "$1" ]] ; then
+        mv "$1" "$2"
+    else
+        echo "$1 gone; will not modify"
+    fi
+}
+rmdir_if_exists Music
+rmdir_if_exists Public
+rmdir_if_exists Templates
+rmdir_if_exists Videos
+
+mvdir_if_exists Desktop desktop
+mvdir_if_exists Downloads downloads
+mvdir_if_exists Documents work
+
+user_dirs_contents='
+# This file is written by xdg-user-dirs-update
+# If you want to change or add directories, just edit the line you'\''re
+# interested in. All local changes will be retained on the next run
+# Format is XDG_xxx_DIR="$HOME/yyy", where yyy is a shell-escaped
+# homedir-relative path, or XDG_xxx_DIR="/yyy", where /yyy is an
+# absolute path. No other format is supported.
+# 
+XDG_DESKTOP_DIR="$HOME/desktop"
+XDG_DOWNLOAD_DIR="$HOME/downloads"
+XDG_TEMPLATES_DIR="$HOME/work"
+XDG_PUBLICSHARE_DIR="$HOME/public_html"
+XDG_DOCUMENTS_DIR="$HOME/work"
+XDG_MUSIC_DIR="$HOME"
+XDG_PICTURES_DIR="$HOME"
+XDG_VIDEOS_DIR="$HOME"
+'
+echo "$user_dirs_contents" > ~/.config/user-dirs.dirs
+
+touch .bootstrapped
+
+.local/bin/fish
